Move validation_tasks under validation
