+++ app/views/planning_applications/validation/tasks/_items_counter.html.erb
