+++ app/views/planning_applications/validation/tasks/_redact_documents.html.erb
