+++ spec/system/planning_applications/validating/check_legislation_spec.rb
@@ -15,7 +15,7 @@ RSpec.describe "Check legislation" do
       planning_application.application_type.update(part: 1, section: "A")
 
       sign_in assessor
-      visit "/planning_applications/#{planning_application.id}/validation_tasks"
+      visit "/planning_applications/#{planning_application.id}/validation/tasks"
       click_link("Check legislative requirements")
     end
 
@@ -88,7 +88,7 @@ RSpec.describe "Check legislation" do
 
     before do
       sign_in assessor
-      visit "/planning_applications/#{planning_application.id}/validation_tasks"
+      visit "/planning_applications/#{planning_application.id}/validation/tasks"
     end
 
     it "does not show the check legislation tasklist" do
